refactor: isolate Menu stories
