+++ packages/react-aria-components/stories/index.stories.tsx
@@ -15,7 +15,7 @@ import {Button, Calendar, CalendarCell, CalendarGrid, Cell, Checkbox, Column, Co
 import {classNames} from '@react-spectrum/utils';
 import clsx from 'clsx';
 import {FocusRing, isTextDropItem, mergeProps, useButton, useClipboard, useDrag} from 'react-aria';
-import {MyListBoxItem} from './utils';
+import {MyListBoxItem, MyMenuItem} from './utils';
 import React, {useRef, useState} from 'react';
 import {RouterProvider} from '@react-aria/utils';
 import styles from '../example/index.css';
@@ -175,55 +175,6 @@ export const SelectRenderProps = () => (
       </>
     )}
   </Select>
-);
-
-export const MenuExample = () => (
-  <MenuTrigger>
-    <Button aria-label="Menu">☰</Button>
-    <Popover>
-      <Menu className={styles.menu} onAction={action('onAction')}>
-        <Section className={styles.group}>
-          <Header style={{fontSize: '1.2em'}}>Section 1</Header>
-          <MyMenuItem>Foo</MyMenuItem>
-          <MyMenuItem>Bar</MyMenuItem>
-          <MyMenuItem>Baz</MyMenuItem>
-          <MyMenuItem href="https://google.com">Google</MyMenuItem>
-        </Section>
-        <Separator style={{borderTop: '1px solid gray', margin: '2px 5px'}} />
-        <Section className={styles.group}>
-          <Header style={{fontSize: '1.2em'}}>Section 2</Header>
-          <MyMenuItem>Foo</MyMenuItem>
-          <MyMenuItem>Bar</MyMenuItem>
-          <MyMenuItem>Baz</MyMenuItem>
-        </Section>
-      </Menu>
-    </Popover>
-  </MenuTrigger>
-);
-
-export const MenuComplex = () => (
-  <MenuTrigger>
-    <Button aria-label="Menu">☰</Button>
-    <Popover>
-      <Menu className={styles.menu}>
-        <MyMenuItem>
-          <Text slot="label">Copy</Text>
-          <Text slot="description">Description</Text>
-          <Keyboard>⌘C</Keyboard>
-        </MyMenuItem>
-        <MyMenuItem>
-          <Text slot="label">Cut</Text>
-          <Text slot="description">Description</Text>
-          <Keyboard>⌘X</Keyboard>
-        </MyMenuItem>
-        <MyMenuItem>
-          <Text slot="label">Paste</Text>
-          <Text slot="description">Description</Text>
-          <Keyboard>⌘V</Keyboard>
-        </MyMenuItem>
-      </Menu>
-    </Popover>
-  </MenuTrigger>
 );
 
 export const NumberFieldExample = () => (
@@ -804,17 +755,6 @@ function MyColumn(props) {
         </div>
       )}
     </Column>
-  );
-}
-
-function MyMenuItem(props) {
-  return (
-    <MenuItem
-      {...props}
-      className={({isFocused, isSelected}) => classNames(styles, 'item', {
-        focused: isFocused,
-        selected: isSelected
-      })} />
   );
 }
